+++ src/components/masters/programs-list/index.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as ProgramsListPage } from './ProgramsListPage';
+export { default as ProgramListPage } from './ProgramListPage';
