+++ src/pages/index.jsx
@@ -3,7 +3,7 @@ import { StaticQuery, graphql } from 'gatsby';
 import { IntlProvider } from 'react-intl';
 
 import withAuthentication from '../components/common/withAuthentication';
-import { ProgramsListPage } from '../components/masters/programs-list';
+import { ProgramListPage } from '../components/masters/programs-list';
 
 const UserProgramsQuery = graphql`
   query {
@@ -27,7 +27,7 @@ const IndexPage = props => (
     <StaticQuery
       query={UserProgramsQuery}
       render={data => (
-        <ProgramsListPage programQueryData={data.allSitePage.edges} {...props} />
+        <ProgramListPage programQueryData={data.allSitePage.edges} {...props} />
       )}
     />
   </IntlProvider>
